Client External Data Fetcher 리턴 타입 수정 (#13)
* client모듈 external data fetch 인터페이스 작성(가안)

* feat: house 생성, 수정, 삭제 기능 추가

* temp: 매물 데이터 생성 배치 추가 (가안)

* client모듈 external data fetch 세부 구현 적용
오피스텔도 반영

* zigbang 원룸과 오피스텔의 factory 분리

* temp: 작업중

* client모듈 피드백 반영

* fix: gradle 관련 설정 수정

* client모듈 피드백 반영

* client모듈 build.gradle.kts수정 & data fetcher throw던지게 수정

* refractor: minor fix

* temp

* external data fetcher return type json -> class로 변경 초안

* 논의대로 단순하게 일단 Json으로 원복
JsonObject는 역직렬화가 안돼서 JsonNode로 타입 변경
Jackson이 Default이길래 Jackson사용

* fetcher에서 null type 에러 리턴하게 변경

* feat: house 정보 create/update 기본 배치 추가

* Fetcher의 return type을 JsonNode -> ExternalData계열 interface로 변경

* Fetcher의 return type을 JsonNode -> ExternalData계열 interface로 변경

* client모듈 인터페이스화

---------

+++ house/src/main/kotlin/xyz/bukinator/house/model/embeddable/Price.kt
@@ -18,7 +18,7 @@ class Price(
 
     @Column(name = "price_manage", nullable = false)
     @Comment("관리세")
-    var priceManage: Int,
+    var priceManage: Double,
 
     @Column(name = "price_manage_includes")
     @Convert(converter = StringToListConverter::class)
